backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
duckdb==0.10.0
python-multipart==0.0.6
python-dotenv==1.0.0
